feat: change the delete blog operation
Now deleting a blog is possible only if the token sent with the request is the same as that of the blog's creator

+++ blog-list/controllers/blogs.js
@@ -40,8 +40,18 @@ blogsRouter.get('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }else {
+    const blog = await Blog.findById(request.params.id)
+    if (blog.user.toString() === decodedToken.id) {
+      await Blog.findByIdAndDelete(request.params.id)
+      response.status(204).end()
+    }else {
+      return response.status(401).json({ error: 'user isn\'t the creator of the blog' })
+    }
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
